# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/users/invites/{id}/approve": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Принимает приглашение в группу по его ID. Пользователь добавляется в участники группы.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users inf"
                ],
                "summary": "Принять приглашение в группу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID приглашения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Приглашение успешно принято",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный ID или ошибка обработки приглашения (например, уже обработано или нет доступа)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/users/invites/{id}/reject": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отклоняет приглашение в группу по его ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users inf"
                ],
                "summary": "Отклонить приглашение в группу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID приглашения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Приглашение успешно отклонено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный ID или ошибка обработки приглашения (например, уже обработано или нет доступа)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users/notifications/viewed": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Помечает указанное уведомление как просмотренное для текущего пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users inf"
                ],
                "summary": "Отметить уведомление как просмотренное",
                "parameters": [
                    {
                        "description": "ID уведомления для отметки",
                        "name": "notification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Уведомление успешно отмечено как просмотренное",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или уведомление не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/users/notify": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список непросмотренных уведомлений и ожидающих приглашений в группы для текущего пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users inf"
                ],
                "summary": "Получить уведомления пользователя",
                "responses": {
                    "200": {
                        "description": "Список уведомлений и приглашений",
                        "schema": {
                            "$ref": "#/definitions/services.GetNotifyResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "services.GetNotifyResponse": {
            "type": "object",
            "properties": {
                "invites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.InviteDTO"
                    }
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.NotificationDTO"
                    }
                }
            }
        },
...
        "services.InviteDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "groupId": {
                    "type": "integer"
                },
                "groupName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "services.NotificationDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "sendAt": {
                    "type": "string"
                },
                "sent": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "viewed": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer \u003cyour_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}